Rename stress-test to repo-sync-test
This test really isn't a stress-test as it is a test to see how much
load we can send to the server before we exceed a maximum repo sync
time.
