+++ package.json
@@ -33,10 +33,10 @@
     "test": "NODE_ENV=test mocha 'test/**/*.test.ts' --timeout 20000",
     "posttest": "nyc report --reporter=cobertura --reporter=html",
     "test.post-report": "open ./coverage/index.html ./coverage/mochawesome.html",
+    "test.repo-sync": "node -r sucrase/register src/bin/repo-sync-test/index.ts",
+    "test.repo-sync.console": "LOG_LEVEL=trace node -r sucrase/register src/bin/repo-sync-test/index.ts | node -r sucrase/register src/bin/repo-sync-test/console.ts",
+    "test.repo-sync.continuous": "node -r sucrase/register src/bin/repo-sync-test/continuous.ts | pino-pretty",
     "test.report": "yarn test.with-report && yarn test.post-report",
-    "test.stress": "node -r sucrase/register src/bin/stress-test/index.ts",
-    "test.stress.console": "LOG_LEVEL=trace node -r sucrase/register src/bin/stress-test/index.ts | node -r sucrase/register src/bin/stress-test/console.ts",
-    "test.stress.continuous": "node -r sucrase/register src/bin/continuous-stress-test/index.ts | pino-pretty",
     "test.watch": "yarn test --watch --watch-files 'src/**/*, test/**/*'",
     "test.with-report": "nyc yarn test --reporter mocha-multi-reporters --reporter-options configFile=./mocha-config.json"
   },
