--- pm2/continuous-stress-test.json
@@ -1,11 +1,0 @@
-{
-  "apps": [
-    {
-      "name": "continuous-stress-test",
-      "script": "lib/bin/continuous-stress-test/index.js",
-      "error_file": "logs/continuous-stress-test/error.log",
-      "out_file": "logs/continuous-stress-test/out.log",
-      "node_args": "--max-old-space-size=16384"
-    }
-  ]
-}
